Fix MakeModelCommand , add features tu Model and queryBuilder

+++ lucid/Commands/MakeModelCommand.php
@@ -3,6 +3,8 @@ namespace Lucid\Commands;
 
 class MakeModelCommand {
     public function handle(string $modelName): void {
+        $tableName = strtolower($modelName) . 's';
+
         $modelTemplate = <<<PHP
 <?php
 namespace App\Models;
@@ -10,11 +12,25 @@ namespace App\Models;
 use Lucid\Model;
 
 class $modelName extends Model {
-    protected static string \$table = strtolower('$modelName') . 's';
+    protected static string \$table = '$tableName';
 }
 PHP;
 
-        $filePath = __DIR__ . '/../../app/Models/' . $modelName . '.php';
+        $directory = __DIR__ . '/../../app/Models/';
+        $filePath = $directory . $modelName . '.php';
+
+        // Crear la carpeta si no existe
+        if (!is_dir($directory)) {
+            mkdir($directory, 0755, true);
+           // echo "Directorio app/Models creado.\n";
+        }
+
+        // Evitar sobrescribir si ya existe
+        if (file_exists($filePath)) {
+            echo "El modelo $modelName ya existe. Operación cancelada.\n";
+            return;
+        }
+
         file_put_contents($filePath, $modelTemplate);
         echo "Modelo $modelName creado en app/Models.\n";
     }
